refactor: Card -> PokeCard, and create a generic `Card` UI component
